requirements.txt
-i https://pypi.org/simple
boto3==1.26.142
botocore==1.29.142 ; python_version >= '3.7'
certifi==2023.5.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
exceptiongroup==1.1.1 ; python_version < '3.11'
idna==3.4 ; python_version >= '3.5'
iniconfig==2.0.0 ; python_version >= '3.7'
jmespath==1.0.1 ; python_version >= '3.7'
packaging==23.1 ; python_version >= '3.7'
pluggy==1.0.0 ; python_version >= '3.6'
pytelegrambotapi==4.12.0
pytest==7.3.1
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.31.0 ; python_version >= '3.7'
s3transfer==0.6.1 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tomli==2.0.1 ; python_version < '3.11'
urllib3==1.26.16 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
